+++ R/RRPP.support.code.r
@@ -308,12 +308,17 @@ lm.args.from.formula <- function(cl){
   Dy <- NULL
   Y <- try(eval(lm.args$formula[[2]], lm.args$data, parent.frame()),
            silent = TRUE)
+  nms <- rownames(lm.args$data)
   
   if(inherits(Y, "try-error"))
     stop("Data are missing from either the data frame or global environment.\n", 
          call. = FALSE)
   
-  if(is.vector(Y)) Y <- as.matrix(Y)
+  if(is.vector(Y)) {
+    Y <- as.matrix(Y)
+    if(!is.null(nms)) rownames(Y) <- nms
+    Dy <- NULL
+  }
   
   if(is.matrix(Y) || is.data.frame(Y)) {
     if(isSymmetric(Y)) {
@@ -321,18 +326,13 @@ lm.args.from.formula <- function(cl){
       if(any(Dy < 0)) stop("Distances in distance matrix cannot be less than 0\n",
                            call. = FALSE)
       lm.args$formula <- update(lm.args$formula, Y ~ .)
-    }
+    } else Dy <- NULL
   }
   
   if(inherits(Y, "dist")) {
     if(any(Y < 0)) stop("Distances in distance matrix cannot be less than 0")
     Dy <- Y
     Y <- pcoa(Y)
-  }
-  
-  if(is.vector(Y)) {
-    Y <- as.matrix(Y)
-    Dy <- NULL
   }
   
   if(is.array(Y) && length(dim(Y)) > 2) 
@@ -349,32 +349,40 @@ lm.args.from.formula <- function(cl){
   
   n <- NROW(Y)
   
-  if(!is.null(lm.args$data)) lm.args$data <- makeDF(form, lm.args$data, n)
+  if(!is.null(lm.args$data)) {
+    lm.args$data <- makeDF(form, lm.args$data, n)
+    rownames(lm.args$data) <- nms
+  }
+  
+  
   if(is.null(lm.args$data)) {
     lm.args$data <- list()
     lm.args$data$Y <- as.matrix(Y)
+    rownames(lm.args$data$Y) <- nms
   }
   
   f <- try(do.call(lm, lm.args), silent = TRUE)
   
   if(inherits(f, "try-error")) {
+    nms <- rownames(Y)
     lm.args$data$Y <- as.matrix(Y)
+    rownames(lm.args$data$Y) <- nms
     f <- try(do.call(lm, lm.args), silent = TRUE)
   }
   
   if(inherits(f, "try-error")) 
-      stop("Independent variables are missing from either the data frame or 
+    stop("Independent variables are missing from either the data frame or 
            global environment,\n", 
-           call. = FALSE)
-
+         call. = FALSE)
+  
   out <- list(Terms = f$terms, model = f$model, 
-       Y = as.matrix(f$y))
+              Y = as.matrix(f$y))
   if(!is.null(Dy)) {
     d <- as.matrix(Dy)
     if(nrow(d) != NROW(out$Y)) d <- d[rownames(Y), rownames(Y)]
     out$D <- as.dist(d)
   }
-
+  
   out
 }