+++ NEWS.md
@@ -9,6 +9,7 @@
 ### BUG FIXES
 * Fixed missing coefficients for GLS estimation in `coef.lm.rrpp`
 * Fixed missing data issue in `lm.rrpp` (passed onto `lm`).
+* Fixed some `as.matrix` names dropping in support code.
 
 # CHANGES IN RRPP 1.0.0 (Major Release)