+++ NEWS.md
@@ -1,3 +1,16 @@
+
+# CHANGES IN RRPP 1.0.0.9000
+
+### NEW FEATURES
+* `na.omit.rrpp.data.frame` added for handling missing data.
+
+### OTHER CHANGES
+
+### BUG FIXES
+* Fixed missing coefficients for GLS estimation in `coef.lm.rrpp`
+* Fixed missing data issue in `lm.rrpp` (passed onto `lm`).
+* Fixed some `as.matrix` names dropping in support code.
+
 # CHANGES IN RRPP 1.0.0 (Major Release)
 
 ### NEW FEATURES